requirements.txt
streamlit>=1.28.0
requests>=2.31.0
youtube-transcript-api>=0.6.1
reportlab>=4.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11 